+++ litellm-proxy-extras/litellm_proxy_extras/utils.py
@@ -7,8 +7,7 @@ import subprocess
 import tempfile
 import time
 from pathlib import Path
-from types import MappingProxyType
-from typing import Final, Optional
+from typing import Optional
 
 from litellm_proxy_extras._logging import logger
 from litellm_proxy_extras.replica_identity import (
@@ -50,17 +49,6 @@ _SPEND_LOGS_PK_CLAUSE_RE = re.compile(
     r'|ADD\s+(?:CONSTRAINT\s+"[^"]*"\s+)?PRIMARY\s+KEY\s*\([^)]*\))$',
     re.IGNORECASE,
 )
-
-_PRISMA_ATTEMPTS: Final = 4
-
-_TRANSIENT_PRISMA_FAILURES: Final = MappingProxyType(
-    {
-        "deadlock detected": "a deadlock on the migration advisory lock (a concurrent migrate deploy)",
-        "P1001": "an unreachable database server",
-        "P1002": "a database server that timed out",
-    }
-)
-
 
 PARTITIONED_SPEND_LOGS_PUSH_ERROR = (
     "LiteLLM_SpendLogs is a partitioned table (see db_scripts/partition_spend_logs.sql), "
@@ -284,23 +272,6 @@ class ProxyExtrasDBManager:
             check=True,
             capture_output=True,
             env=prisma_env,
-        )
-
-    @staticmethod
-    def _transient_prisma_failure(stderr: str) -> str | None:
-        """Why a failed prisma command is worth retrying, or None.
-
-        v1 retried every failure, so it absorbed a database that was not up yet
-        or another instance holding the migration lock. v2 fails fast, which is
-        right for a broken migration and wrong for these.
-        """
-        return next(
-            (
-                reason
-                for marker, reason in _TRANSIENT_PRISMA_FAILURES.items()
-                if marker in stderr
-            ),
-            None,
         )
 
     @staticmethod
@@ -684,7 +655,7 @@ class ProxyExtrasDBManager:
     @staticmethod
     def _setup_database_v2(use_migrate: bool) -> bool:
         """
-        v2 migration resolver (what the proxy CLI selects by default).
+        v2 migration resolver (opt-in via --use_v2_migration_resolver).
 
         Runs `prisma migrate deploy` and handles standard recovery paths
         (P3005 baseline, P3009/P3018 idempotent errors). Critically, it does
@@ -705,46 +676,20 @@ class ProxyExtrasDBManager:
             original_dir = os.getcwd()
             os.chdir(migrations_dir)
             try:
-                for attempt in range(_PRISMA_ATTEMPTS):
-                    try:
-                        subprocess.run(
-                            [_get_prisma_command(), "db", "push", "--accept-data-loss"],
-                            timeout=prisma_command_timeout(),
-                            check=True,
-                            capture_output=True,
-                            text=True,
-                            env=_get_prisma_env(),
-                        )
-                        return True
-                    except subprocess.TimeoutExpired:
-                        logger.info(
-                            "prisma db push attempt %s timed out, retrying",
-                            attempt + 1,
-                        )
-                        time.sleep(random.randrange(5, 15))
-                    except subprocess.CalledProcessError as e:
-                        stderr = e.stderr or ""
-                        transient = ProxyExtrasDBManager._transient_prisma_failure(
-                            stderr
-                        )
-                        # Re-raise as RuntimeError so proxy_cli.py's
-                        # `except RuntimeError` catches it and exits cleanly.
-                        if transient is None or attempt == _PRISMA_ATTEMPTS - 1:
-                            raise RuntimeError(
-                                f"prisma db push failed.\n\nDetail: {e}"
-                                f"\n\nPrisma error:\n{stderr}"
-                            ) from e
-                        logger.info(
-                            "prisma db push attempt %s failed on %s, retrying. "
-                            "Prisma error:\n%s",
-                            attempt + 1,
-                            transient,
-                            stderr,
-                        )
-                        time.sleep(random.randrange(5, 15))
-                raise RuntimeError(
-                    f"prisma db push failed after {_PRISMA_ATTEMPTS} attempts."
+                subprocess.run(
+                    [_get_prisma_command(), "db", "push", "--accept-data-loss"],
+                    timeout=prisma_command_timeout(),
+                    check=True,
+                    env=_get_prisma_env(),
                 )
+                return True
+            except (
+                subprocess.CalledProcessError,
+                subprocess.TimeoutExpired,
+            ) as e:
+                # Re-raise as RuntimeError so proxy_cli.py's
+                # `except RuntimeError` catches it and exits cleanly.
+                raise RuntimeError(f"prisma db push failed.\n\nDetail: {e}") from e
             finally:
                 os.chdir(original_dir)
 
@@ -754,7 +699,7 @@ class ProxyExtrasDBManager:
         original_dir = os.getcwd()
         os.chdir(migrations_dir)
         try:
-            for attempt in range(_PRISMA_ATTEMPTS):
+            for attempt in range(4):
                 try:
                     result = subprocess.run(
                         [_get_prisma_command(), "migrate", "deploy"],
@@ -869,36 +814,16 @@ class ProxyExtrasDBManager:
                             f"Manual intervention required.\n\nPrisma error:\n{stderr}"
                         ) from e
 
-                    transient = ProxyExtrasDBManager._transient_prisma_failure(stderr)
-                    if transient is None:
-                        raise RuntimeError(
-                            "Database migration failed and cannot be auto-recovered. "
-                            f"Manual intervention required.\n\nPrisma error:\n{stderr}"
-                        ) from e
-
-                    if attempt == _PRISMA_ATTEMPTS - 1:
-                        raise RuntimeError(
-                            f"Database migration failed after "
-                            f"{_PRISMA_ATTEMPTS} attempts on {transient}. "
-                            "Check database connectivity and load."
-                            f"\n\nPrisma error:\n{stderr}"
-                        ) from e
-
-                    logger.info(
-                        "prisma migrate deploy attempt %s failed on %s, retrying. "
-                        "Prisma error:\n%s",
-                        attempt + 1,
-                        transient,
-                        stderr,
-                    )
-                    time.sleep(random.randrange(5, 15))
-                    continue
+                    raise RuntimeError(
+                        "Database migration failed and cannot be auto-recovered. "
+                        f"Manual intervention required.\n\nPrisma error:\n{stderr}"
+                    ) from e
 
             raise RuntimeError(
-                f"Database migration failed after {_PRISMA_ATTEMPTS} "
-                "attempts (retry loop exhausted by timeouts or repeated "
-                "idempotent-recovery continues). Check database connectivity, "
-                "load, and _prisma_migrations ledger state."
+                "Database migration failed after 4 attempts (retry loop "
+                "exhausted by timeouts or repeated idempotent-recovery "
+                "continues). Check database connectivity, load, and "
+                "_prisma_migrations ledger state."
             )
         finally:
             os.chdir(original_dir)
@@ -946,11 +871,10 @@ class ProxyExtrasDBManager:
 
         Args:
             use_migrate: Whether to use prisma migrate instead of db push
-            use_v2_resolver: Run the v2 migration resolver (safer during
+            use_v2_resolver: Opt into the v2 migration resolver (safer during
                 rolling deploys; does not run the diff-and-force recovery
-                that causes schema thrashing). Defaults to False here so
-                direct callers keep the old behavior; the proxy CLI passes
-                True, so the proxy's runtime default is v2.
+                that causes schema thrashing). Defaults to False for
+                backwards compatibility.
 
         Returns:
             bool: True if setup was successful, False otherwise
@@ -968,7 +892,7 @@ class ProxyExtrasDBManager:
     @staticmethod
     def _run_migrations(use_migrate: bool, use_v2_resolver: bool) -> bool:
         if use_v2_resolver:
-            logger.info("Using v2 migration resolver")
+            logger.info("Using v2 migration resolver (--use_v2_migration_resolver)")
             return ProxyExtrasDBManager._setup_database_v2(use_migrate=use_migrate)
 
         schema_path = ProxyExtrasDBManager._get_prisma_dir() + "/schema.prisma"
