+++ litellm/proxy/proxy_cli.py
@@ -913,14 +913,13 @@ class ProxyInitializationHelpers:
     envvar="ENFORCE_PRISMA_MIGRATION_CHECK",
 )
 @click.option(
-    "--use_v2_migration_resolver/--use_legacy_migration_resolver",
-    default=True,
+    "--use_v2_migration_resolver",
+    is_flag=True,
+    default=False,
     help=(
-        "Which database migration resolver to run at startup. The default v2 "
-        "resolver avoids the diff-and-force recovery path that can cause schema "
-        "thrashing during rolling deploys where two LiteLLM versions contend for "
-        "the same DB. Pass --use_legacy_migration_resolver, or set "
-        "USE_V2_MIGRATION_RESOLVER=false, to fall back to v1."
+        "Opt into the v2 migration resolver. Avoids the diff-and-force recovery "
+        "path that can cause schema thrashing during rolling deploys where two "
+        "LiteLLM versions contend for the same DB. Default is the v1 resolver."
     ),
     envvar="USE_V2_MIGRATION_RESOLVER",
 )
@@ -1311,11 +1310,10 @@ def run_server(
                 else:
                     if not use_v2_migration_resolver:
                         print(
-                            "\033[1;33mLiteLLM Proxy: Using the legacy (v1) migration resolver. "
-                            "The default v2 resolver is safer: it avoids the diff-and-force "
-                            "recovery that caused schema thrashing during rolling deploys. "
-                            "Remove --use_legacy_migration_resolver / "
-                            "USE_V2_MIGRATION_RESOLVER=false to switch back to it.\033[0m"
+                            "\033[1;33mLiteLLM Proxy: Using default (v1) migration resolver. "
+                            "If your deployment has seen schema thrashing during rolling "
+                            "deploys, try --use_v2_migration_resolver (safer: avoids the "
+                            "diff-and-force recovery that caused the thrash).\033[0m"
                         )
                     try:
                         setup_ok: Final = PrismaManager.setup_database(
@@ -1323,10 +1321,10 @@ def run_server(
                             use_v2_resolver=use_v2_migration_resolver,
                         )
                     except RuntimeError as e:
-                        # Raised on unrecoverable migration errors: permission
-                        # failures from either resolver, the v2 resolver's
-                        # non-idempotent failures, and any `prisma db push`
-                        # against a partitioned LiteLLM_SpendLogs.
+                        # Raised on unrecoverable migration errors: the v2
+                        # resolver's non-idempotent failures and permission
+                        # issues, and any `prisma db push` against a
+                        # partitioned LiteLLM_SpendLogs.
                         print(
                             f"\033[1;31mLiteLLM Proxy: Database migration cannot proceed. {e}\033[0m",
                             file=sys.stderr,
